break up utils into 4 separate modules
