+++ lib/browser/common/util.js
@@ -1,3 +1,4 @@
+export * from './util/misc'
 export * from './util/check'
 export * from './util/dom-helpers'
 export * from './util/tag-helpers'
