+++ lib/browser/common/util/check.js
@@ -3,8 +3,28 @@ import {
   T_OBJECT,
   T_FUNCTION,
   T_UNDEF,
+  RE_SVG_TAGS,
+  RE_BOOL_ATTRS,
   RE_RESERVED_NAMES
 } from './../global-variables'
+
+/**
+ * Check whether a DOM node must be considered a part of an svg document
+ * @param   { String } name -
+ * @returns { Boolean } -
+ */
+export function isSVGTag(name) {
+  return RE_SVG_TAGS.test(name)
+}
+
+/**
+ * Check Check if the passed argument is undefined
+ * @param   { String } value -
+ * @returns { Boolean } -
+ */
+export function isBoolAttr(value) {
+  return RE_BOOL_ATTRS.test(value)
+}
 
 /**
  * Check if passed argument is a function
@@ -79,24 +99,4 @@ export function isWritable(obj, key) {
  */
 export function isReservedName(value) {
   return RE_RESERVED_NAMES.test(value)
-}
-
-/**
- * Check whether an array contains an item
- * @param   { Array } array - target array
- * @param   { * } item - item to test
- * @returns { Boolean } -
- */
-export function contains(array, item) {
-  return ~array.indexOf(item)
-}
-
-/**
- * Faster String startsWith alternative
- * @param   { String } str - source string
- * @param   { String } value - test string
- * @returns { Boolean } -
- */
-export function startsWith(str, value) {
-  return str.slice(0, value.length) === value
 }
