+++ lib/browser/common/util/dom-helpers.js
@@ -1,30 +1,5 @@
-import {
-  RE_SVG_TAGS,
-  RE_HTML_ATTRS,
-  RE_BOOL_ATTRS
-} from './../global-variables'
-
-import {
-  isUndefined
-} from './check'
-
-/**
- * Check whether a DOM node must be considered a part of an svg document
- * @param   { String } name -
- * @returns { Boolean } -
- */
-export function isSVGTag(name) {
-  return RE_SVG_TAGS.test(name)
-}
-
-/**
- * Check Check if the passed argument is undefined
- * @param   { String } name -
- * @returns { Boolean } -
- */
-export function isBoolAttr(value) {
-  return RE_BOOL_ATTRS.test(value)
-}
+import { RE_HTML_ATTRS } from './../global-variables'
+import { isUndefined } from './check'
 
 /**
  * Create a generic DOM node
@@ -110,4 +85,43 @@ export function walkAttrs(html, fn) {
   let m
   while (m = RE_HTML_ATTRS.exec(html))
     fn(m[1].toLowerCase(), m[2] || m[3] || m[4])
+}
+
+/**
+ * Walk down recursively all the children tags starting dom node
+ * @param   { Object }   dom - starting node where we will start the recursion
+ * @param   { Function } fn - callback to transform the child node just found
+ * @param   { Object }   context - fn can optionally return an object, which is passed to children
+ */
+export function walkNodes(dom, fn, context) {
+  if (dom) {
+    const res = fn(dom, context)
+    let next
+    // stop the recursion
+    if (res === false)
+      return
+    else {
+      dom = dom.firstChild
+
+      while (dom) {
+        next = dom.nextSibling
+        walkNodes(dom, fn, res)
+        dom = next
+      }
+    }
+  }
+}
+
+/**
+ * Check whether a DOM node is in stub mode, useful for the riot 'if' directive
+ * @param   { Object }  dom - DOM node we want to parse
+ * @returns { Boolean } -
+ */
+export function isInStub(dom) {
+  while (dom) {
+    if (dom.inStub)
+      return true
+    dom = dom.parentNode
+  }
+  return false
 }
