+++ lib/browser/common/util/tag-helpers.js
@@ -3,9 +3,12 @@ import {
 } from './dom-helpers'
 
 import {
-  isArray,
-  isWritable,
-  contains
+  contains,
+  each
+} from './misc'
+
+import {
+  isArray
 } from './check'
 
 import {
@@ -19,33 +22,6 @@ import {
 
 import Tag from './../../tag/tag'
 import { tmpl } from 'riot-tmpl'
-/**
- * Specialized function for looping an array-like collection with `each={}`
- * @param   { Array } els - collection of items
- * @param   {Function} fn - callback function
- * @returns { Array } the array looped
- */
-export function each(els, fn) {
-  var len = els ? els.length : 0
-
-  for (var i = 0, el; i < len; i++) {
-    el = els[i]
-    // return false -> current item was removed by fn during the loop
-    if (el != null && fn(el, i) === false) i--
-  }
-  return els
-}
-
-/**
- * Convert a string containing dashes to camel case
- * @param   { String } string - input string
- * @returns { String } my-string -> myString
- */
-export function toCamel(string) {
-  return string.replace(/-(\w)/g, function(_, c) {
-    return c.toUpperCase()
-  })
-}
 
 /**
  * Detect the tag implementation by a DOM node
@@ -135,24 +111,6 @@ export function unmountAll(expressions) {
 }
 
 /**
- * Helper function to set an immutable property
- * @param   { Object } el - object where the new property will be set
- * @param   { String } key - object key where the new property will be stored
- * @param   { * } value - value of the new property
-* @param   { Object } options - set the propery overriding the default options
- * @returns { Object } - the initial object
- */
-export function defineProperty(el, key, value, options) {
-  Object.defineProperty(el, key, extend({
-    value,
-    enumerable: false,
-    writable: false,
-    configurable: true
-  }, options))
-  return el
-}
-
-/**
  * Get the tag name of any DOM node
  * @param   { Object } dom - DOM node we want to parse
  * @param   { Boolean } skipName - hack to ignore the name attribute when attaching to parent
@@ -169,30 +127,6 @@ export function getTagName(dom, skipName) {
 }
 
 /**
- * Extend any object with other properties
- * @param   { Object } src - source object
- * @returns { Object } the resulting extended object
- *
- * var obj = { foo: 'baz' }
- * extend(obj, {bar: 'bar', foo: 'bar'})
- * console.log(obj) => {bar: 'bar', foo: 'bar'}
- *
- */
-export function extend(src) {
-  var obj, args = arguments
-  for (var i = 1; i < args.length; ++i) {
-    if (obj = args[i]) {
-      for (var key in obj) {
-        // check if this property of the source object could be overridden
-        if (isWritable(src, key))
-          src[key] = obj[key]
-      }
-    }
-  }
-  return src
-}
-
-/**
  * With this function we avoid that the internal Tag methods get overridden
  * @param   { Object } data - options we want to use to extend the tag instance
  * @returns { Object } clean object without containing the riot internal reserved words
@@ -206,42 +140,6 @@ export function cleanUpData(data) {
     if (!RE_RESERVED_NAMES.test(key)) o[key] = data[key]
   }
   return o
-}
-
-/**
- * Walk down recursively all the children tags starting dom node
- * @param   { Object }   dom - starting node where we will start the recursion
- * @param   { Function } fn - callback to transform the child node just found
- * @param   { Object }   context - fn can optionally return an object, which is passed to children
- */
-export function walk(dom, fn, context) {
-  if (dom) {
-    var res = fn(dom, context), next
-    // stop the recursion
-    if (res === false) return
-    else {
-      dom = dom.firstChild
-
-      while (dom) {
-        next = dom.nextSibling
-        walk(dom, fn, res)
-        dom = next
-      }
-    }
-  }
-}
-
-/**
- * Check whether a DOM node is in stub mode, useful for the riot 'if' directive
- * @param   { Object }  dom - DOM node we want to parse
- * @returns { Boolean } -
- */
-export function isInStub(dom) {
-  while (dom) {
-    if (dom.inStub) return true
-    dom = dom.parentNode
-  }
-  return false
 }
 
 /**
