+++ lib/browser/tag/parse.js
@@ -2,13 +2,21 @@ import IfExpr from './if'
 import NamedExpr from './named'
 import _each from './each'
 import { tmpl } from 'riot-tmpl'
-import { RIOT_TAG, RE_BOOL_ATTRS } from './../common/global-variables'
-import { walk, getAttr, each, getTag, initChildTag, remAttr } from './../common/util'
+import { RIOT_TAG } from './../common/global-variables'
+
+import {
+  isBoolAttr,
+  walkNodes,
+  getAttr,
+  each,
+  getTag,
+  initChildTag
+} from './../common/util'
 
 export function parseExpressions(root, tag, expressions, includeRoot) {
   var base = {parent: {children: expressions}}
 
-  walk(root, function(dom, ctx) {
+  walkNodes(root, function(dom, ctx) {
     var type = dom.nodeType, parent = ctx.parent, attr, expr, childTag
     if (!includeRoot && dom === root) return {parent: parent}
 
@@ -63,7 +71,7 @@ export function parseExpressions(root, tag, expressions, includeRoot) {
 // it is also passed to fn.
 export function parseAttributes(dom, attrs, tag, fn) {
   each(attrs, function(attr) {
-    var name = attr.name, bool = RE_BOOL_ATTRS.test(name), expr
+    var name = attr.name, bool = isBoolAttr(name), expr
 
     if (name === 'name' || name === 'id') {
       expr = new NamedExpr(dom, name, attr.value, tag)
