+++ lib/browser/tag/tag.js
@@ -8,16 +8,15 @@ import { mixin } from './core'
 
 import {
   GLOBAL_MIXIN,
-  T_FUNCTION,
-  T_UNDEF,
-  T_STRING,
-  RE_RESERVED_NAMES,
   __VIRTUAL_DOM,
   RIOT_TAG_IS,
   RIOT_TAG
 } from './../common/global-variables'
 
 import {
+  isReservedName,
+  isUndefined,
+  isString,
   isFunction,
   isObject,
   isWritable,
@@ -29,7 +28,6 @@ import {
   each,
   walkAttrs,
   toCamel,
-  isInStub,
   setAttr,
   remAttr,
   unmountAll,
@@ -95,7 +93,7 @@ export default function Tag(impl, conf, innerHTML) {
 
   function normalizeData(data) {
     for (var key in item) {
-      if (typeof self[key] !== T_UNDEF && isWritable(self, key))
+      if (!isUndefined(self[key]) && isWritable(self, key))
         self[key] = data[key]
     }
   }
@@ -103,8 +101,8 @@ export default function Tag(impl, conf, innerHTML) {
   function inheritFromParent () {
     each(Object.keys(self.parent), function(k) {
       // some properties must be always in sync with the parent tag
-      var mustSync = !RE_RESERVED_NAMES.test(k) && contains(propsInSyncWithParent, k)
-      if (typeof self[k] === T_UNDEF || mustSync) {
+      var mustSync = !isReservedName(k) && contains(propsInSyncWithParent, k)
+      if (isUndefined(self[k]) || mustSync) {
         // track the property to keep in sync
         // so we can keep it updated
         if (!mustSync) propsInSyncWithParent.push(k)
@@ -147,7 +145,7 @@ export default function Tag(impl, conf, innerHTML) {
       var props = []
       var obj
 
-      mix = typeof mix === T_STRING ? mixin(mix) : mix
+      mix = isString(mix) ? mixin(mix) : mix
 
       // check if the mixin is a function
       if (isFunction(mix)) {
@@ -162,7 +160,7 @@ export default function Tag(impl, conf, innerHTML) {
       // loop the keys in the function prototype or the all object keys
       each(props, function(key) {
         // bind methods to self (if not init or riot.observable functions)
-        if (!/^(init|on|off|one|trigger)$/.test(key)) {
+        if (!isReservedName(key)) {
           const descriptor = Object.getOwnPropertyDescriptor(instance, key)
           const hasGetterSetter = descriptor && (descriptor.get || descriptor.set)
 
@@ -228,7 +226,6 @@ export default function Tag(impl, conf, innerHTML) {
       while (dom.firstChild) root.appendChild(dom.firstChild)
       if (root.stub) self.root = root = parent.root
     }
-
 
     defineProperty(self, 'root', root)
     self.isMounted = true
